+++ src/main/java/com/dragonminez/common/stats/StatsData.java
@@ -560,9 +560,9 @@ public class StatsData {
 		return totalCost;
 	}
 
-	public int calculateStatIncrease(int maxStatsToAdd, int availableTPs, int maxStats) {
+	public int calculateStatIncrease(int maxStatsToAdd, long availableTPs, int maxStats) {
 		int statsIncreased = 0;
-		int costAccumulated = 0;
+		long costAccumulated = 0;
 		int currentTotalStats = stats.getTotalStats();
 
 		while (statsIncreased < maxStatsToAdd) {
